modify steemapi relocation files
