+++ src/main/resources/properties/Readme.txt
@@ -1,0 +1,7 @@
+You need to make "steem.properties" file in this path.
+
+<< steem.properties content >>
+account.name=steem_account
+account.posting=private postingkey
+account.active=private activekey
+account.memo=private memokey
